move redux provider to longview only
