+++ packages/manager/src/components/AbuseTicketBanner/AbuseTicketBanner.utils.ts
@@ -1,4 +1,4 @@
-import type { Notification } from '@linode/api-v4/lib/account';
+import type { Notification } from '@linode/api-v4';
 
 export const getAbuseTickets = (notifications: Notification[]) => {
   return notifications.filter(
