+++ packages/manager/src/dev-tools/DevTools.tsx
@@ -5,7 +5,6 @@ import { styled } from '@mui/material';
 import { QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import React from 'react';
-import { Provider } from 'react-redux';
 
 import { Draggable } from './components/Draggable';
 import { DesignTokensTool } from './DesignTokensTool';
@@ -16,7 +15,6 @@ import { ServiceWorkerTool } from './ServiceWorkerTool';
 import { isMSWEnabled } from './utils';
 
 import type { QueryClient } from '@tanstack/react-query';
-import type { ApplicationStore } from 'src/store';
 
 export type DevToolsView = 'design-tokens' | 'mocks' | 'react-query';
 
@@ -28,7 +26,6 @@ const reactQueryDevtoolsStyle = {
 
 interface Props {
   queryClient: QueryClient;
-  store: ApplicationStore;
 }
 
 export const DevTools = (props: Props) => {
@@ -90,103 +87,101 @@ export const DevTools = (props: Props) => {
   }, []);
 
   return (
-    <Provider store={props.store}>
-      <Draggable draggable={isDraggable}>
-        <div
-          className={`dev-tools ${isMSWEnabled ? 'dev-tools--msw' : ''} ${
-            isOpen ? 'dev-tools--open' : ''
-          } ${isOpen && isDraggable ? 'isDraggable' : ''}
-          `.trim()}
-        >
-          {!isDraggable && (
-            <div className="dev-tools__toggle">
-              <button
-                className="dev-tools-button"
-                onClick={() => setIsOpen(!isOpen)}
-              >
-                <Handyman />
-              </button>
-            </div>
-          )}
-          {isOpen && (
-            <div className="dev-tools__draggable-toggle">
-              <button
-                className="dev-tools-button"
-                onClick={handleDraggableToggle}
-                title="Toggle draggable"
-              >
-                {isDraggable ? <CloseIcon /> : <OpenInNewIcon />}
-              </button>
-            </div>
-          )}
-          <div className="dev-tools__body">
-            <div className="dev-tools__content">
-              <div className="dev-tools__status-bar">
-                <div>
-                  <EnvironmentToggleTool />
-                </div>
-                <div className="dev-tools__segmented-button">
-                  <button
-                    className={`toggle-button dev-tools-button ${
-                      view === 'mocks' && 'toggle-button--on'
-                    }`}
-                    onClick={handleOpenMocks}
-                  >
-                    Mocks
-                  </button>
-                  <button
-                    className={`toggle-button dev-tools-button ${
-                      view === 'react-query' && 'toggle-button--on'
-                    }`}
-                    onClick={handleOpenReactQuery}
-                  >
-                    React Query
-                  </button>
-                  <button
-                    className={`toggle-button dev-tools-button ${
-                      view === 'design-tokens' && 'toggle-button--on'
-                    }`}
-                    onClick={handleOpenDesignTokens}
-                  >
-                    Design Tokens
-                  </button>
-                </div>
-                <div>
-                  <button
-                    className="dev-tools-button"
-                    onClick={handleGoToPreferences}
-                  >
-                    Go to Preferences
-                  </button>
-                </div>
+    <Draggable draggable={isDraggable}>
+      <div
+        className={`dev-tools ${isMSWEnabled ? 'dev-tools--msw' : ''} ${
+          isOpen ? 'dev-tools--open' : ''
+        } ${isOpen && isDraggable ? 'isDraggable' : ''}
+        `.trim()}
+      >
+        {!isDraggable && (
+          <div className="dev-tools__toggle">
+            <button
+              className="dev-tools-button"
+              onClick={() => setIsOpen(!isOpen)}
+            >
+              <Handyman />
+            </button>
+          </div>
+        )}
+        {isOpen && (
+          <div className="dev-tools__draggable-toggle">
+            <button
+              className="dev-tools-button"
+              onClick={handleDraggableToggle}
+              title="Toggle draggable"
+            >
+              {isDraggable ? <CloseIcon /> : <OpenInNewIcon />}
+            </button>
+          </div>
+        )}
+        <div className="dev-tools__body">
+          <div className="dev-tools__content">
+            <div className="dev-tools__status-bar">
+              <div>
+                <EnvironmentToggleTool />
               </div>
-              <div className="dev-tools__main" ref={devToolsMainRef}>
-                {view === 'mocks' && (
-                  <>
-                    <div className="dev-tools__main__column">
-                      <FeatureFlagTool />
-                    </div>
-                    <div className="dev-tools__main__column">
-                      <ServiceWorkerTool />
-                    </div>
-                  </>
-                )}
-                {view === 'react-query' && (
-                  <StyledReactQueryDevtoolsContainer
-                    style={reactQueryDevtoolsStyle}
-                  >
-                    <QueryClientProvider client={props.queryClient}>
-                      <ReactQueryDevtools initialIsOpen={true} />
-                    </QueryClientProvider>
-                  </StyledReactQueryDevtoolsContainer>
-                )}
-                {view === 'design-tokens' && <DesignTokensTool />}
+              <div className="dev-tools__segmented-button">
+                <button
+                  className={`toggle-button dev-tools-button ${
+                    view === 'mocks' && 'toggle-button--on'
+                  }`}
+                  onClick={handleOpenMocks}
+                >
+                  Mocks
+                </button>
+                <button
+                  className={`toggle-button dev-tools-button ${
+                    view === 'react-query' && 'toggle-button--on'
+                  }`}
+                  onClick={handleOpenReactQuery}
+                >
+                  React Query
+                </button>
+                <button
+                  className={`toggle-button dev-tools-button ${
+                    view === 'design-tokens' && 'toggle-button--on'
+                  }`}
+                  onClick={handleOpenDesignTokens}
+                >
+                  Design Tokens
+                </button>
               </div>
+              <div>
+                <button
+                  className="dev-tools-button"
+                  onClick={handleGoToPreferences}
+                >
+                  Go to Preferences
+                </button>
+              </div>
+            </div>
+            <div className="dev-tools__main" ref={devToolsMainRef}>
+              {view === 'mocks' && (
+                <>
+                  <div className="dev-tools__main__column">
+                    <FeatureFlagTool />
+                  </div>
+                  <div className="dev-tools__main__column">
+                    <ServiceWorkerTool />
+                  </div>
+                </>
+              )}
+              {view === 'react-query' && (
+                <StyledReactQueryDevtoolsContainer
+                  style={reactQueryDevtoolsStyle}
+                >
+                  <QueryClientProvider client={props.queryClient}>
+                    <ReactQueryDevtools initialIsOpen={true} />
+                  </QueryClientProvider>
+                </StyledReactQueryDevtoolsContainer>
+              )}
+              {view === 'design-tokens' && <DesignTokensTool />}
             </div>
           </div>
         </div>
-      </Draggable>
-    </Provider>
+      </div>
+    </Draggable>
   );
 };
