+++ packages/manager/src/features/Longview/store/longview/longview.events.ts
@@ -2,7 +2,7 @@ import { getAllLongviewClients } from './longview.requests';
 
 import type { EventStatus } from '@linode/api-v4/lib/account';
 import type { Dispatch } from 'redux';
-import type { EventHandler } from 'src/store/types';
+import type { EventHandler } from 'src/features/Longview/store/types';
 
 const longviewEventHandler: EventHandler = (event, dispatch, getState) => {
   const { action, entity, status } = event;
