--- packages/manager/src/hooks/types.ts
@@ -1,5 +1,0 @@
-import type { AnyAction } from 'redux';
-import type { ThunkDispatch } from 'redux-thunk';
-import type { ApplicationState } from 'src/store';
-
-export type Dispatch = ThunkDispatch<ApplicationState, undefined, AnyAction>;
