+++ packages/manager/src/index.tsx
@@ -3,13 +3,11 @@ import CssBaseline from '@mui/material/CssBaseline';
 import { QueryClientProvider } from '@tanstack/react-query';
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import { Provider as ReduxStoreProvider } from 'react-redux';
 
 import { CookieWarning } from 'src/components/CookieWarning';
 import 'src/exceptionReporting';
 import { SplashScreen } from 'src/components/SplashScreen';
 import { setupInterceptors } from 'src/request';
-import { storeFactory } from 'src/store';
 
 import './index.css';
 import { App } from './App';
@@ -17,7 +15,6 @@ import { ENABLE_DEV_TOOLS } from './constants';
 import { LinodeThemeWrapper } from './LinodeThemeWrapper';
 
 const queryClient = queryClientFactory('longLived');
-const store = storeFactory();
 
 setupInterceptors();
 
@@ -27,16 +24,14 @@ const Main = () => {
   }
 
   return (
-    <ReduxStoreProvider store={store}>
-      <QueryClientProvider client={queryClient}>
-        <LinodeThemeWrapper>
-          <CssBaseline enableColorScheme />
-          <React.Suspense fallback={<SplashScreen />}>
-            <App />
-          </React.Suspense>
-        </LinodeThemeWrapper>
-      </QueryClientProvider>
-    </ReduxStoreProvider>
+    <QueryClientProvider client={queryClient}>
+      <LinodeThemeWrapper>
+        <CssBaseline enableColorScheme />
+        <React.Suspense fallback={<SplashScreen />}>
+          <App />
+        </React.Suspense>
+      </LinodeThemeWrapper>
+    </QueryClientProvider>
   );
 };
 
@@ -53,7 +48,7 @@ async function loadApp() {
 
     const root = createRoot(devToolsRootContainer);
 
-    root.render(<DevTools queryClient={queryClient} store={store} />);
+    root.render(<DevTools queryClient={queryClient} />);
   }
 
   const container = document.getElementById('root');
